statistic page optimised

+++ app/Http/Controllers/AdminNewsController.php
@@ -26,6 +26,7 @@ class AdminNewsController extends NewsController
 
     public function edit(News $news)
     {
+        abort_if(!$news->id, 404);
         return view('admin.editNews', [
             'news' => $news,
         ]);
@@ -33,12 +34,14 @@ class AdminNewsController extends NewsController
 
     public function update(Request $request, News $news)
     {
+        abort_if(!$news->id, 404);
         $newNews = $this->updateFunction($request, $news);
         return redirect("/news/$newNews->slug");
     }
 
     public function destroy(News $news)
     {
+        abort_if(!$news->id, 404);
         \Mail::to($news->author->email)->send(new \App\Mail\ArticleDeleted($news));
         $news->delete();
         flash('Статья успешно удалена', 'warning');
